refactor the mapper to make it easy to handle
